requirements.txt
webp==0.1.3
pillow==8.4.0
requests==2.25.11
aiohttp==3.7.4.post0
